refactor(command): move external package to messaging
