+++ internal/core/command/controller/messaging/external_test.go
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package external
+package messaging
 
 import (
 	"context"
